feat: Implement TestDefaultConfig for pallet-evm
